Create OAuth component
